backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data validation and settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client for LLM API calls
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.98.3

# Code quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
